fix: addresses mmr `find_peaks` bug (#5182)
Description
---
Currently, the Tari MMR repo contains a bug on the `find_peaks` function, in `base_layer/mmr/src/common.rs`. The source of the problem arises from the fact that the current implementation does not consider possible Merkle Mountain Ranges, in which all nodes are filled to the possible top (that is, every possible spawned peak exists in the MMR). Moreover, the current function implementation handles this issue by returning an empty `vec![]`. This works as a 'silent error' being passed upstream. In order to handle this problem, we impose that the height to check that a new peak exists only decreases if we are guaranteed that every peak at the same height does in fact not exists. In this way, proofs can be created for every MMR size. 

PS: Want to add further unit tests.

Motivation and Context
---

How Has This Been Tested?
---
More extensive unit testing.


<!-- Does this include a breaking change? If so, include this line as a footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a database, resync the chain -->