Correct miner logging
I think the log was meant to print to stdout as well as the miner
appender. It was done by duplicated a section of yaml, but that infact
overrides one for the other. So we only had stdout, and never to the log
file.

Remove the duplicated block, and have all output go to stdout and the
miner logfile.