feat(consensus)!: add balanced binary merkle tree (#5189)
Description
---
Implement balanced binary merkle tree and use it instead of mmr for vns. It has no effect on the base layer functionality apart for being a breaking change (it changes the vn merkle root).  In the base layer only the generation is used. The proofs are used only in `tari-dan` (signing votes).
The advantages of MMR over binary tree is the time complexity when adding an element. But that's not our case. We regenerate the tree everytime. So it's much simpler to generate balanced binary tree. Which is the best option for this. The proof time/size for binary tree is always log n. For MMR the worst case scenario is 2*log n (when there is peak of every possible height). 
The merged proofs are part of another PR.

Some comparison between balanced binary merkle tree vs merkle mountain range.
The data set I used was number of VNs of 1000 to 1M (step by 1000), committee size from 10 to 200 (step by 10) (In reality the committee can be bigger, this is representing the number of proofs).
For generating the proofs the random subset of VNs were selected (same for BMT and MMR). 

- **Tree generation takes 62.03% of the time**
![tree_generation](https://user-images.githubusercontent.com/35243812/220345863-3a717a26-6abd-43ab-a1a9-7bfefd369016.gif)
- **Generating proofs takes 42.37% of the time**
**Generating merged proofs takes 85.96% of the time** 
In the graph below the green represents only the merging time. But in reality we need to generate the proof anyway and then merge them (figure above takes this into account)
![proof_generation](https://user-images.githubusercontent.com/35243812/220347909-f8a1f0e8-5af4-40e9-b57b-fa03b284d3d5.gif)
- **Verifying proofs takes 68.98% of the time**
**Verifying merged proofs takes 54.77% of the time**
![verifying_proofs](https://user-images.githubusercontent.com/35243812/220345871-f83b9ca1-744e-4b7e-81f3-d3774c1c910a.gif)
- **Proof size is 76.12% of the size**
**Merged proof size is 9.47% of the size**
![proof_size](https://user-images.githubusercontent.com/35243812/220345861-e0830ffe-c4b6-4a65-bc48-3c5be2203048.gif)


How Has This Been Tested?
---
There are couple cargo tests.
<!-- Does this include a breaking change? If so, include this line as a footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a database, resync the chain -->
BREAKING CHANGE:  **The Merkle root for VNs is different.** It's equal in the genesis block, but once there is a single vn registration it starts to differ.