docs: add second-preimage warning to MMR documentation (#5208)
Description
---
Updates documentation for the Merkle mountain range implementation to add a warning about second-preimage resistance.

Motivation and Context
---
The Merkle mountain range implementation can be instantiated in a way that makes it [susceptible](https://github.com/tari-project/tari/issues/5206) to second-preimage attacks, such that a given root does not computationally bind to a set of leaf nodes.

The linked issue discusses possible mitigations, with varying degrees of complexity. This PR implements the simplest approach, which is to warn the user that the caller is responsible for arranging its leaf node and MMR hashers to mitigate this attack risk.

It may still be desirable to implement one of the other mitigations discussed in the linked issue.

How Has This Been Tested?
---
No testing is needed.