fix: windows path format in log4rs files
log4rs expects the paths to be in a unix format regardless of the system
it's running on. When we started doing the dynamic injecting of the base
path windows systems injected using a windows format, which mixed with
the unix format. Resulting in strings that looked liked this:
`C:\\brianp\.tari\config/log/base-node/network.log`

when it log4rs wants the final format to be like this:
`C://brianp/.tari/config/log/base-node/network.log`

This commit does a naive character replacement on the string path just
before we do the path injection.

Co-authored-by: hansieodendaal 39146854+hansieodendaal@users.noreply.github.com