Use a generic DbConnection where possible
Have the sqlite common crate provide a generic db connection struct for
use in simplified cases such as testing, and hopefully we could move all
or most connection types over to this in the future.

Moves the error type of PooledDbConnection from the contacts crate over
to sqlite_common as it's more appropriate in the common crate.