feat(wallet)!: use ECDH shard secret for burn mask with claim pubkey (#5238)
Description
---
Derives and uses domain-separated ECDH shared secret for burnt output commitment mask and encrypted value.

Motivation and Context
---
Previously, the DAN layer can claim the burnt output and deposit it into an account. However, the mask and value were known only to the L1 wallet, meaning that the claimed funds could not be spent. The PR uses an ECDH shared secret and allows the DAN wallet to derive the shared secret in the same way that it does for any other confidential transfer.

The console wallet will generate $H(k.A)$ where $A$ is the claim key provided by the user and $k$ is a blinding factor generated using the console wallet's seed key. Usually, this will be the key that owns the account component (ref: https://github.com/tari-project/tari-dan/pull/425). The public blinding factor (`reciprocal_claim_public_key`) is returned to the `burn_tari` caller and is passed to the `ClaimBurn` instruction, which in turn, 'mints' the <commitment, reciprocal_claim_public_key, encrypted_value> ConfidentailResource tuple. 

NOTE: Burns UTXOs _are not_ recoverable because the wallet cannot determine the mask/decrypt the value without the original claim_public_key which is not stored on-chain.

How Has This Been Tested?
---
Existing cucumber test for burn execute part of the code modified in this PR, however test coverage needs to be improved

<!-- Does this include a breaking change? If so, include this line as a footer -->
BREAKING CHANGE: `burn_tari` GRPC response changed (re-add proto to POSTman if you use that).