feat!: add claim public key to OutputFeatures (#5239)
Description
---
- Adds confidential claim data to chain when burn_tari is invoked with
claim_public_key set.
- Updates output recovery in wallet to recover burnt outputs

Motivation and Context
---
Claim key is required for mask and value recovery of burnt funds on L2.
Unfortunately, since burnt outputs are not part of the utxo set, the
wallet does not receive these outputs when recovering. Resolving this is
out of scope for this PR.

How Has This Been Tested?
---
Manually: normal wallet recovery OK. Resync base node OK.

<!-- Does this include a breaking change? If so, include this line as a
footer -->
BREAKING CHANGE: Not database or network breaking, but a chain split
will occur if a burnt utxo with the claim public key set is broadcast
and some nodes have not upgraded.