feat: improved passphrase flow (#5279)
Description
---
Improves the flow for setting or changing a passphrase.

Closes [issue 5127](https://github.com/tari-project/tari/issues/5127).

Motivation and Context
---
When setting or
[changing](https://github.com/tari-project/tari/pull/5175) a wallet
passphrase, the console wallet provides
[feedback](https://github.com/tari-project/tari/pull/5111) on the
strength of the provided passphrase. In the case of a weak passphrase,
it does not prompt the user to choose a better one.

This PR implements a better flow for this process, as shown in [this
flowchart](https://github.com/tari-project/tari/issues/5127#issuecomment-1487027081).

How Has This Been Tested?
---
Tested manually.

What process can a PR reviewer use to test or verify this change?
---
Testing needs to be done manually to assert that the process represented
by the linked flowchart is implemented. Manual testing should cover the
entire flow for these two operations:
- Setting the passphrase for a new wallet
- Changing the passphrase for an existing wallet

Breaking Changes
---
None.