feat!: change signature construction to allow better HW support (#5282)
Description
---
Changes the challenges of all signatures to not be
`Hash(Key||Nonce||Feat_1||Feat_2||..||Feat_x) `but rather
`Hash(Key||Nonce||Hash(Feat_1||Feat_2||..||Feat_x))`

Adds the version to the kernel signature as this is currently not
checked.

Motivation and Context
---
This is done because hardware devices need to sign all the data they
need to be sent over to the hardware device to be signed. If we have to
send all the features over this becomes quite significant in bytes. This
is now changed to only send the `message` field as a single 32-byte
array and this still comments to all the features.

The kernel version is technically malleable as the version is not
committed to in the signature, this is now fixed.
How Has This Been Tested?
---
All tests pass

What process can a PR reviewer use to test or verify this change?
---
Make sure all chain signatures are changed. 
Adds version to Kernel signature

Breaking Changes
---
Change all chain signature challenges: Kernel, Script, Metadata

- [x] None
- [x] Requires data directory on base node to be deleted
- [x] Requires hard fork
- [x] Other - Please specify
Complete chain reset. 
<!-- Does this include a breaking change? If so, include this line as a
footer -->

BREAKING CHANGE: Complete chain reset