fix: resize transaction tab windows
PR #4942 added a row to the tx detail view, but did not increase the
size of the box. As a result, the maturity is cut off.

This PR makes the detail box one row wider, and adjusts the minimum size
of the tx box to be 8 (this is the absolute minimum -- accounting for
headers and one each of pending and completed txs).

Fixes #5289