feat!: move key manager service to key_manager (#5284)
Description
---
Moves the key_manager_services to key_manager_create

Motivation and Context
---
Currently, all transactions are built by supplying private keys to the
constructor. The wallet implements a Key_manager_service that safely
constructs the private key but this is then passed around in code to
where it is used to sign transactions etc.

We want to refactor this, that the code only ask for a signature and not
a private key. This means the private key generation should happen where
core can access it. This will be refactored in a future PR. This is part
of allowing support of Hardware devices in the future.

How Has This Been Tested?
---
Unit tests and manual tx sending



Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Please specify
Requires data directory to be deleted on the wallet

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->
BREAKING CHANGE: Requires data directory to be deleted on the wallet
because it changes how the wallet database structure looks.