feat: feature gates (#5287)
Description
---
This PR is for reflective purposes regarding feature gating at compile
time. See https://github.com/tari-project/tari/issues/5135 for more
info. This PR as it stands is to be used as an example of a network
_type_ based feature gate at compile time.
It allows setting an ENVVAR for the network with conditional compilation
for features. The feature sets can be imported across any crate in the
project easily and then used with the `#[cfg(tari_feature_...)]`
attribute macro.

I've also introduced a secondary commit for binaries to perform a quick
network check against itself to ensure the intended network is supported
by the binary. This check is made non-invasive (not conditionally
compiling possible network) as that method of check ends up being much
more tedious.

Motivation and Context
---
We want some features on some networks, but maybe not all.

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
Compile a binary with a set network
`TARI_NETWORK=nextnet cargo build --release --bin tari_base_node`

then run the binary with a test network as a parameter:

`./tari_base_node --network esme` 

and receive an error:

```
The application exited because of an internal network error. The network esmeralda is invalid for this binary built for NextNet
```

*Please note*

That running the binary will cause the build to default to the test
network _always_. Meaning this will not work:

```
$ TARI_NETWORK=nextnet cargo build --release --bin tari_base_node
$ cargo run --release --bin tari_base_node --network nextnet
```

The second command `cargo run` will re-build the binary, and without the
`TARI_NETWORK` env set the binary will default to a test binary.

In development if you want to call run and connect to a non test network
you need to also pass the TARI_NETWORK envvar instead of the `--network`
flag.
`TARI_NETWORK=nextnet cargo run --bin tari_base_node --release --
--network nextnet`


Breaking Changes
---

- [x] None

---------

Co-authored-by: Cayle Sharrock <CjS77@users.noreply.github.com>
Co-authored-by: SW van Heerden <swvheerden@gmail.com>