refactor: cucumber directory (#5337)
Description
---
- This is a cleanup and restructuring of the integration tests.
- Moving steps into separate step files
- Moving integration code from the test utils dir into the main
integration crate

Motivation and Context
---
cucumber.rs was becoming unwieldy. We had utils in the main crate that
was completely unused. And all the code existed in the test dir.

How Has This Been Tested?
---
With cucumber.

What process can a PR reviewer use to test or verify this change?
---
See if you like the new structure or not. 
See if the cucumber tests pass.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->