chore: update location of compiled bins in s3 (#5358)
Description
---
Separate the bins by network type to make it easier to distinguish on
the tari website for download.


Motivation and Context
---
It makes it easier to parse on the website, allowing us to deploy the
new design which should help people disambiguate which binary they
should download.


How Has This Been Tested?
---
#yolo

What process can a PR reviewer use to test or verify this change?
---
Know bash better than I do and validate with 100% certainty the change
is most excellent

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->