fix: added missing log4rs features (#5356)
Description
---
Added missing log4rs features to applications which use these features.

Motivation and Context
---
Base node panics because the sample log config uses a number of log4rs
features that are not compiled in.
The dependency was not added to the common lib as this adds extra
dependencies to a lot of libs and it breaks wasm build.

How Has This Been Tested?
---
Base node runs

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->