fix: fix panic when no public addresses (#5367)
Description
---
If there are no public addresses specified, there was a panic. This is
solved by not starting a liveness service if there is no public address.

NOTE: I have changed the reference on tari-crypto to be the lowest
version that it works with (i.e. 0.16, instead of 0.16.*). This is the
best practice for libraries, according to Rust, and also helps with
mismatches on the tari-dan repo.

Motivation and Context
---
When using TCP, a node might not have a public address, this may cause a
panic with the current code. AFAIK a comms node (base node, wallet, dan
indexer) can still work without a public address, but it can only
initiate outbound connections.

How Has This Been Tested?
---
existing tests

What process can a PR reviewer use to test or verify this change?
---
TODO: I'm busy checking this on the DAN indexer, which has no public
address during testing

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->