docs: remove an outdated comment (#5387)
Description
---
Removes an outdated comment, in what is possibly the silliest and
smallest PR.

Motivation and Context
---
A comment was outdated.

How Has This Been Tested?
---
It hasn't.

What process can a PR reviewer use to test or verify this change?
---
Look for the comment. Find that it has vanished. Contemplate the
consequences.