fix: loop on mismatched passphrase entry (#5396)
Description
---
If the user enters a new or changed passphrase and fails to confirm it,
an exit error is immediately returned. This PR updates to prompt the
user to try again, up to a sanity limit.

Closes [issue 5391](https://github.com/tari-project/tari/issues/5391).

Motivation and Context
---
The user is prompted to enter and confirm a new passphrase in two cases:
when creating a new wallet, and when changing the passphrase of an
existing wallet. In both cases, failure to confirm the new passphrase
correctly returns an exit error.

This PR prompts the user again if the new passphrase and its
confirmation do not match.

We probably don't want the user to get stuck in an infinite loop and go
insane, so there is a sanity limit of three tries. If the user fails
this many times, we return the exit error.

How Has This Been Tested?
---
Tested manually.

What process can a PR reviewer use to test or verify this change?
---
Test the following scenarios manually:
- Create a new wallet. Fail to confirm a new passphrase three times.
Confirm that the process exits.
- Create a new wallet. Confirm a new passphrase within three tries.
Confirm that the process succeeds.
- Change an existing wallet's passphrase. Fail to confirm a new
passphrase three times. Confirm that the process exits.
- Change an existing wallet's passphrase. Confirm a new passphrase
within three tries. Confirm that the process succeeds.