Keep the World Client struct
A logger was added to print the world client struct after a failure has
occured. This was showing most collection as empty due to the fact we
were purposfully clearing it after each singular test run. The purpose
of clearing it was to ensure no crossover or spillage from one test to
another but this isn't a real problem with cucumber. The world object is
cloned new into every test run and isn't shared, so we don't need to
clean it out.