Maintain the logs of a single test run
Right now the logs will continue to ammend a single file. Sometimes this
is bothersome when wanting to debug the last one, as the file may be
quite large. This commit will move the logs into the temp directory of
the last run isolating that runs logs for easier debugging.