test: improvements for debugging integration tests (#5402)
Description
---
This PR does a small handful of refactoring to help when debugging the
integration tests:
- Better use of temporary directories for base node data
- Print base node identity files
- Better use of temporary directories for wallet data
- Move logs from each test run into the last run directory for easier to
search and more isolated logs
- Fix 2 features so they'll run now
- Stop using in memory databases for DHT, it's possible for their
reference to be dropped and the DHT tables to be wiped causing weird
behaviour in the tests
- Stop manual cleaning of the World after a test runs, there is no need
to do this, and it produced undesired behavior when printing panic data
- Miner was attempting to connect to Esme, ensure it stays on the
localnetwork

Motivation and Context
---
The suite can be difficult to debug, hopefully this makes it easier

How Has This Been Tested?
---
Locally

What process can a PR reviewer use to test or verify this change?
---
See if it passes CI

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->