feat: improved base node monitoring (#5390)
Add the ability to run tokio_console on the console_wallet. 

I found one case where the UI would use a lot of CPU, but I couldn't
find the cause.

Reduced the UI caching time to be more responsive.

Renamed base_node_monitor_refresh_interval to
base_node_monitor_max_refresh_interval and implemented a backoff
strategy.
If a new block is found when querying the base node, immediately query
it again, backing off each time it has the same block. This allows you
to update quickly if the node is syncing. Otherwise, backoff to the max
interval. The default is 90 seconds.
(was previously 3 seconds)

Changed output and transaction manager service to rely on the base node
service to send them an event when a new block is encountered. Both were
doing this check themselves in addition to the check the base node
service was doing.

The base node service also only checked on the height, now the hash is
also checked.

---------

Co-authored-by: Hansie Odendaal <39146854+hansieodendaal@users.noreply.github.com>