test: fix wallet ffi base dirs in tests to prevent ci from returning false failures (#5418)
Description
---
The wallet ffi started using the log directory for it's data directory.
This conflicted with the new code to move logs into the test run dir.

Motivation and Context
---
CI was passing and reporting failure.

How Has This Been Tested?
---
Locally

What process can a PR reviewer use to test or verify this change?
---
Watch CI pass

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->