feat: refactor configuration for chat so ffi can create and accept a config file (#5426)
Description
---
This refactoring come on the heels of making testing easier via the chat
configuration. As a result we also solve the problem about peer seeds in
FFI. Previously the ffi required passing in a collection of peer seeds.
This wasn't in any way ideal. Now we take a standard tari configuration
setup, which will utilize the DNS seeds in the default configuration.

Motivation and Context
---
ChatFFI had a bad way to connect to peers. It needed fixing. Also
cleanup of the integration tests was desired and making configuration
simpler is helpful.

How Has This Been Tested?
---
Locally and CI

No new test was added for the configuration creation yet. As that config
defaults to real config, not local config.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->